+++ src/components/App.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import './App.css';
+import '../assets/App.css';
 
 function App() {
   return (
@@ -7,8 +7,8 @@ function App() {
       <header className="App-header">
         <p>
           Edit
-          <code>src/App.js</code>
-          sand save to reload.
+          <code> src/App.js </code>
+          and save to reload.
         </p>
         <a
           className="App-link"
